v0.0.4 - Support multiple images in docker testing (#3)
**git-shared-checkout**
  * v0.0.5f
  * Update references
  * Added --remove
  * Improved usage
  * Support --list
  * fix overlooked typos

**Building [build-funcs]**:
  * fix: Ensure link removed when cleaning
  * Permit multiple docker-images
  * Summarise times if --stats chosen
  * setupPython automatically does --dev on development branches
  * fix: Bug that disabled --clean
  * issue: Build failure on root not recoverable with --clean
  * chore: Shuffle ='all' dump for more pleasing order
  * chore: Shuffle clean so that the directories are deleted first - faster\!
  * chore: Removed pointless print 'Didn't touch '.'
  * chore: Cleaner logging
  * Permit running in a single docker
  * feat: Cleaning workspace


**Testing [test-funcs]**:
  * Import more tests from external sources

**do-run-in-docker**:
* Release  v0.0.1
* Improved help
* Improved summary
* image_with_apt()]: Add python to basic image
* fix: Correct mistake in help

* chore: Shuffle to _internalUse
* chore: Add support for 'THIS_DIR_REL'
* chore: Remove unused file

* fix: Avoid ././fname summary



**Other** :
* do-install-bashlibs.sh -> install/do-install-bashlibs.sh
* Improved link display
* Update Colours (Add RED)
* Update _quoteIfNeeded
* Improved test support
* Cleanup
* Clean help + Add more deprecation warnings
* Improved deprecation checks
* Shuffle source including:
    - Move git-shared-checkout to own part
     - Creating _sample-projects

* Add .pre-commit, .gitignore etc
* fix: Missed return values
* displayPath(): Improve message when link is invalid

+++ lib-app.inc.bash
@@ -1,8 +1,27 @@
+# shellcheck shell=bash
+# shellcheck disable=SC2317
 
 BUILD_FUNCS_DIR="$(dirname "$(realpath -m "${BASH_SOURCE[0]}")")"
 source "${BUILD_FUNCS_DIR%/}/lib-common.inc.bash"
 
-#!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+APP_PARAMS=("$@")
+
+function exitWithVersion()
+{
+    echo "$CMD_AS_DISPLAY ${APP_VERSION:-v?.?.?}"
+    exit 0
+}
+
+if [[ -n "${APP_VERSION:-}" ]]; then
+    # Do outside of all the wrappers so that it is always available, even if there is a problem with the wrapper setup
+    for arg in "$@"; do
+        [[ "$arg" == '--version' ]] && exitWithVersion
+    done
+fi
+
+
+#==============================================================================================
 #
 # This is a standard template for bash scripts that uses:
 #  * APP_VERSION     | Optional but recommended
@@ -86,7 +105,8 @@ function giveCodeHelp()
     echo  "    ─────────────────────┼─────────────────────────────────────────────────┼────────────────────────"
     echo "    \$EXE_NAME            │ The name of the executable (e.g. 'myscript.sh') │ ${EXE_NAME:-❓  Not defined}"
     echo "    \$THIS_EXE            │ The full path to the executable                 │ ${THIS_EXE:-❓  Not defined}"
-    echo "    \$THIS_DIR            │ The directory of the executable                 │ ${THIS_DIR:-❓  Not defined}"
+    echo "    \$EXE_DIR             │ The directory of the executable                 │ ${EXE_DIR:-❓  Not defined}"
+    echo "    \$PROJ_DIR            │ The root directory of the project               │ ${PROJ_DIR:-❓  Not defined}"
     echo "    \$THIS_EXE_AS_DISPLAY │ The display path of the executable              │ ${THIS_EXE_AS_DISPLAY:-❓  Not defined}"
     echo "    \$CMD_AS_DISPLAY      │ The executable in a form that can be run        │ ${CMD_AS_DISPLAY:-❓  Not defined}"
     echo ""
@@ -151,7 +171,7 @@ function do_exitWithHelp()
                 echo -e "     --install  : Installs to $_installDirShow "
             fi
         fi
-        echo -e "     --colours=no|yes|auto  (Default 'auto')"
+        echo -e "     --colours=no|yes   (Default: yes)" # |auto  (Default 'auto')"
         echo ""
         echo "Runs with shared checkout:"
         echo -n "     •  "
@@ -165,7 +185,141 @@ function do_exitWithHelp()
         exit 0
     fi
 }
+function do_withOptionalTiming()
+{
+    local with_timing="$1"
+    shift 1
+    local result=0
 
+    if [[ "${with_timing:-}" == 'yes' ]]; then
+        TIMEFORMAT=$'\n-- Total Time Summary ---\nReal:  %3Rs\nUser:  %3Us\nSys:   %3Ss\nCPU:   %P%%\n----------------------'
+        time "$@" || result="$?"
+    else
+        "$@" || result="$?"
+    fi
+    # |!!>| echo "!!!!!!!!! do_withOptionalTiming[$*] : Result: $result" >&2
+    return "$result"
+}
+
+function bashlibs_warn_on_version_if_needed()
+{
+    if [[ -n "${UKKO_BASHLIBS_REF_FORCE:-}" ]] && [[ "${UKKO_BASHLIBS_REF_FORCE:-}" != "${UKKO_BASHLIBS_REF_PREFERRED:-}" ]] ; then
+        echo -n "⚠️  UKKO_BASHLIBS_REF_FORCE=$UKKO_BASHLIBS_REF_FORCE"
+        [[ -n "${UKKO_BASHLIBS_REF_PREFERRED:-}" ]] && echo -n " (Preferred: '${UKKO_BASHLIBS_REF_PREFERRED:-}')"
+        echo ""
+    fi
+}
+
+
+
+# Return with no argument if name starts with '!'
+#
+# param_choose_from_list "exit" "$1" "$2" fred mary tom
+# ---------------|-----------------------|--------------
+#  selected_name | Params $1,$2          | Result
+# ---------------|-----------------------|--------------
+#   value        | --value= fred           -> true
+#   value        | --value= mary           -> true
+#   value        | --value= tom            -> true
+#  !value        | --value= fred           -> true
+#  !value        | --value= mary           -> true
+#  !value        | --value= tom            -> true
+# ---------------|-----------------------|--------------
+#   value        | --value                 ->      FALSE
+#  !value        | --value                 -> true
+# ---------------|-----------------------|--------------
+#   value        | --value= other          ->      FALSE
+#  !value        | --value= other          ->      FALSE
+#   value        | --value=                ->      FALSE
+#  !value        | --value=                ->      FALSE
+# ---------------|-----------------------|--------------
+function param_choose_from_list()
+{
+    function _param_choose_from_list()
+    {
+        local name_for_comparison="$1"
+        local decorated_name_passed="$2"
+        local value_passed="$3"
+
+        if [[ "$name_for_comparison" == '--'* ]] || [[ "$name_for_comparison" == *'=' ]] ; then
+            local _bad="$name_for_comparison"
+            name_for_comparison="${name_for_comparison#--}"
+            name_for_comparison="${name_for_comparison%=}"
+            echo -e "⚠️  Deprecated - param_choose_####() ${name_for_comparison@Q} rather than ${_bad@Q}" >&2
+        fi
+
+        if [[ "${name_for_comparison}" == '!'* ]] ; then
+            name_for_comparison="${name_for_comparison#'!'}"
+            # Accept default if the name is passed without '=value' (e.g. '--exit' rather than '--exit=yes' or '--exit=no')
+            [[ "$decorated_name_passed" == "--${name_for_comparison}" ]] && return 0
+        fi
+
+        [[ "$decorated_name_passed" == "--${name_for_comparison}=" ]] || return 1
+
+
+        shift 3 || true
+
+        app_load_param_validate_from_list "--${name_for_comparison}=" "${value_passed}" "$@"
+    }
+    local _result=0
+    _param_choose_from_list "$@" || _result=$?
+
+
+    # |Logging| echo "ℹ️  param_choose_from_list($*) : Result: ${_result}" >&2
+    return "$_result"
+}
+
+# Return with no argument if name starts with '!'
+#
+# param_choose_yes_no "exit" "$1" "$2"
+# ---------------|-----------------------|--------------
+#  selected_name | Params $1,$2          | Result
+# ---------------|-----------------------|--------------
+#   exit         | --exit= yes             -> true
+#   exit         | --exit= no              -> true
+#  !exit         | --exit= yes             -> true
+#  !exit         | --exit= no              -> true
+# ---------------|-----------------------|--------------
+#   exit         | --exit                  ->      FALSE
+#  !exit         | --exit                  -> true
+# ---------------|-----------------------|--------------
+#   exit         | --exit= other           ->      FALSE
+#  !exit         | --exit= other           ->      FALSE
+#   exit         | --exit=                 ->      FALSE
+#  !exit         | --exit=                 ->      FALSE
+# ---------------|-----------------------|--------------
+function param_choose_yes_no()
+{
+    param_choose_from_list "$@" "yes" "no"
+}
+
+#
+# ref_name="$1"
+# provided_name="$2"
+# provided_value_txt="$3"
+# min_value
+# max_value
+function param_int()
+{
+    local ref_name="$1"
+    local provided_name="$2"
+    local provided_value_txt="$3"
+    local min_value="${4:-}"
+    local max_value="${5:-}"
+
+    [[ "$ref_name" == "$provided_name" ]] || return 1
+
+    local caption="${!ref_name}[${min_value}…${max_value}]"
+    print_extraVerbose "$caption : Validating ${provided_value_txt@Q}"
+
+    [[ "$provided_value_txt" =~ ^[+-]?[0-9]+$ ]] || do_errorExitWithSuggestion "Invalid value for $caption: ${provided_value_txt@Q} is not an integer"
+
+    local provided_value="$provided_value_txt"
+    [[ -n "$min_value" ]] && [[ "$provided_value" -lt "$min_value" ]] && do_errorExitWithSuggestion "Invalid value for $caption: ${provided_value@Q} is less than minimum allowed value of $min_value"
+    [[ -n "$max_value" ]] && [[ "$provided_value" -gt "$max_value" ]] && do_errorExitWithSuggestion "Invalid value for $caption: ${provided_value@Q} is greater than maximum allowed value of $max_value"
+
+    return 0
+}
 
 function app_load_param_validate_from_list()
 {
@@ -173,12 +327,17 @@ function app_load_param_validate_from_list()
     local value="$2"
 
     shift 2 || true
-    #echo "ℹ️  Validating value for $name$value against expected values: [$(asCsvList "$@")]"
+    print_extraVerbose  "app_load_param_validate_from_list[$name,$value]: ($*) : Validating value for $name$value against expected values: [$(asCsvList "$@")]"
 
     for arg in "$@"; do
         [[ "$arg" == "$value" ]] && return 0
     done
-    do_errorExitWithSuggestion "Invalid value for ${name}<value>.  <value> is ${value@Q}, but expected one of [$(asCsvList "$@")]"
+
+    if [[ "$#" == 1 ]]; then
+        do_errorExitWithSuggestion "Invalid value for $name<value>.  Expected: ${1@Q}, but got ${value@Q}"
+    else
+        do_errorExitWithSuggestion "Invalid value for ${name}<value>.  <value> is ${value@Q}, but expected one of [$(asCsvList "$@")]"
+    fi
 }
 
 function load_params()
@@ -186,8 +345,15 @@ function load_params()
     local am_processing_options='yes'
     function _app_get_param_defaults()
     {
-        declare -F app_load_param_defaults  &> /dev/null  || return 0 ; app_load_param_defaults
+        declare -F app_load_param_defaults  &> /dev/null  || return 0
+        app_load_param_defaults
     }
+    function _app_get_param_review()
+    {
+        declare -F app_params_review  &> /dev/null  || return 0
+        app_params_review
+    }
+
     function _app_get_param_option_name_value()
     {
         declare -F app_load_param_option_name_value  &> /dev/null  && app_load_param_option_name_value "$1" "${2:-}"
@@ -203,26 +369,24 @@ function load_params()
     _app_get_param_defaults
     local giveCodeHelp='no'
     local giveStdHelp='no'
+
     for arg in "$@"; do
         [[ -z "$arg" ]] && continue
         if [[ "$am_processing_options" == "yes" ]] && [[ "$arg" == '-'* ]]; then
             if [[ "$arg" == '--' ]] ; then
                 am_processing_options='no'
             elif [[ "$arg" == '--version' ]] && [[ -n "${APP_VERSION:-}" ]]; then
-                echo "$(basename "$THIS_EXE") $APP_VERSION"
-                exit 0
+                exitWithVersion
             elif [[ "$arg" == '--help' ]]; then
                 giveStdHelp='yes'
             elif [[ "$arg" == '--colours=no' ]] || [[ "$arg" == '--colours=yes' ]] || [[ "$arg" == '--colours=auto' ]]; then
                 colours_load "${arg#--colours=}"
-            elif [[ "${APP_SELF_INSTALL:-}" == "yes" ]] ; then
-                if [[ "$INSTALLATION_NOTE" == "OK" ]] && [[ "$arg" == "--uninstall" ]] ; then
-                    (sleep 0.5 && rm -f "$THIS_EXE" && echo "Removed $THIS_EXE") &
-                    exit 0
-                elif [[ "$arg" == "--install" ]] ; then # Always permit 'install' even if the app is already installed, as this can be used to fix an installation that is not working (e.g. due to missing copy or missing PATH entry)
-                    do_install_directly
-                    exit 0
-                fi
+            elif [[ "${APP_SELF_INSTALL:-}" == "yes" ]] && [[ "$INSTALLATION_NOTE" == "OK" ]] && [[ "$arg" == "--uninstall" ]] ; then
+                (sleep 0.5 && rm -f "$THIS_EXE" && echo "Removed $THIS_EXE") &
+                exit 0
+            elif [[ "${APP_SELF_INSTALL:-}" == "yes" ]] && [[ "$arg" == "--install" ]] ; then # Always permit 'install' even if the app is already installed, as this can be used to fix an installation that is not working (e.g. due to missing copy or missing PATH entry)
+                do_install_directly "      "
+                exit 0
             elif [[ "$arg" == '--code-help' ]] ; then
                 giveCodeHelp='yes'
             elif [[ "$arg" == "--"*"="* ]] ; then
@@ -241,7 +405,7 @@ function load_params()
     elif [[ "$giveStdHelp" == "yes" ]]; then
         do_exitWithHelp ""
     fi
-
+    _app_get_param_review
 }
 
 function do_with_check()
@@ -265,32 +429,32 @@ function do_with_check()
     else
         [[ "$status" == 1 ]] || suffix=" with status $status"
         echo -e "Ran    :${COLOUR[VIVID_RED_STDERR]:-}$caption${COLOUR[OFF_STDERR]:-} -- Failed$suffix" >&2
-        [[ "$caption" == "$*" ]] || echo -e "Command: ${COLOUR[VIVID_BLUE_STDERR]:-}$*${COLOUR[OFF_STDERR]:-}" >&2
+        [[ "$caption" == "$*" ]] || echo -e "Command: ${COLOUR[VIVID_BLUE_STDERR]:-}$(quoteIfNeeded "$@")${COLOUR[OFF_STDERR]:-}" >&2
         exit 1
     fi
 }
 
 
-
 function do_install_directly()
 {
+    local prefix="${1:-}"
     if [[ "$INSTALLATION_NOTE" == "OK" ]]; then
-        echo -e "${COLOUR[VIVID_GREEN_STDOUT]:-}✅ $app_name_and_version is already successfully installed to  ${INSTALL_DIR@Q}${COLOUR[OFF_STDOUT]:-}"
+        echo -e "${prefix}${COLOUR[VIVID_GREEN_STDOUT]:-}✅ $app_name_and_version is already successfully installed to  ${INSTALL_DIR@Q}${COLOUR[OFF_STDOUT]:-}"
         return 0
     fi
 
     if [[ "$INSTALLATION_NOTE" == *"[NotCopied]"* ]] ; then
         if ! do_ensure_file_set  "$INSTALL_DIR_WITH_EXE" "$THIS_EXE" ; then
-            echo "❌  Failed to copy $THIS_EXE to $INSTALL_DIR_WITH_EXE"
+            echo -e "${prefix}${COLOUR[VIVID_RED_STDERR]:-}❌  Failed to copy $THIS_EXE to $INSTALL_DIR_WITH_EXE${COLOUR[OFF_STDERR]:-}"
             return 1
         fi
     fi
 
     if [[ "$INSTALLATION_NOTE" == *"[DirNotInPath]"* ]] ; then
-        echo "⚠️  $app_name_and_version installed to ${INSTALL_DIR@Q} but it is NOT in PATH. Please add ${INSTALL_DIR@Q} to your PATH environment variable." >&2
+        echo -e "${prefix}${COLOUR[VIVID_RED_STDERR]:-}⚠️  $app_name_and_version installed to ${INSTALL_DIR@Q} but it is NOT in PATH. Please add ${INSTALL_DIR@Q} to your PATH environment variable.${COLOUR[OFF_STDERR]:-}" >&2
         return 1
     fi
-     echo "✅  $app_name_and_version installed successfully to ${INSTALL_DIR@Q} and is available in PATH"
+     echo -e "${prefix}✅  ${COLOUR[VIVID_BLUE_STDOUT]:-}$app_name_and_version${COLOUR[OFF_STDOUT]:-} installed successfully to ${INSTALL_DIR@Q} and is available in PATH"
 }
 
 
@@ -375,10 +539,135 @@ function colours_show()
 #################
 #
 
+function app_dumpInfo()
+{
+    local reason="${1:-}"
 
-colours_load 'auto'
+    function dumpEntry_txt()
+    {
+        local name="$1"
+        local value="$2"
 
-# |x| echo "ℹ️  !! RUNNING:  $0 $*" >&2
-load_params "$@"
+        printf " • %-32s = %s\n" "${name}" "${value}"
+    }
 
-app_run "$@"
+    function dumpEntry_var()
+    {
+        local name="$1"
+        shift 1 || true
+
+        local value_txt
+
+        if [[ -v "$name" ]]; then
+            local value="${!name}"
+            value_txt="$(quoteIfNeeded "${value}")"
+        elif [[ $(declare -p "$name" 2>/dev/null) == "declare -a"* ]]; then
+            name="${name}[@]"
+            value_txt="[$(asCsvList "${!name}")]"
+        elif [[ $(declare -p "$name" 2>/dev/null) == "declare -A"* ]]; then
+            value_txt="{"
+            name="${name}[@]"
+            for key in "${!name}" ; do
+                local ref="${name[$key]}"
+                value_txt+=" $(quoteIfNeeded "$key"): $(quoteIfNeeded "${!ref}")"
+            done
+            value_txt+="}"
+        else
+            value_txt="<not set>"
+            return 0
+        fi
+        dumpEntry_txt "\$${name}" "$value_txt" "$*"
+
+    }
+
+
+    {
+        [[ -n "$reason" ]] && echo -e "$reason\n"
+        echo -n "$CMD_AS_DISPLAY"
+        for x in "${APP_PARAMS[@]}" ; do
+            echo -n ' '
+            quoteIfNeeded "$x"
+        done
+        echo ""
+
+        local options=("${!option_@}")
+        for option in "${options[@]}" ; do
+            dumpEntry_var "${option}" "${option@a}"
+        done
+        echo ""
+        dumpEntry_var "APPS_NAME"
+        dumpEntry_var "SUGGEST_HOW_TO_INSTALL_TO_ROOT"
+        dumpEntry_var "VERIFY_ON_BUILD_ENVIRONMENTS"
+        dumpEntry_var "PWD"
+        dumpEntry_var "ORIG_PWD"
+        dumpEntry_var "PROJ_DIR"
+        dumpEntry_var "EXE_DIR"
+        dumpEntry_var "EXE_DIR_AS_DISPLAY"
+        dumpEntry_var "APP_SELF_INSTALL"
+        [[ "${INSTALLATION_NOTE:-[NotCopied]}" == "[NotCopied]" ]] || dumpEntry_var "INSTALLATION_NOTE"
+        dumpEntry_txt "colours"               "${UKKO_COLOURS_CHOSEN@Q}"
+
+    } | withLeftBox >&2
+}
+
+# |!!>| export app_result_fname=''
+# |!!>| echo "app_result_include() - Initialised" >&2
+# |!!>| function app_result_include()
+# |!!>| {
+# |!!>|     set -x
+# |!!>|     local _comment="${2:-}"
+# |!!>|     local _readValue=0
+# |!!>|     echo "!!> ${_comment}: app_result_include($*) : app_result_fname=${app_result_fname@Q}" >&2
+# |!!>|
+# |!!>|     [[ -n "${app_result_fname:-}" ]] && _readValue="$(<"$app_result_fname")"
+# |!!>|     echo "!!> ${_comment}: Loaded value $_readValue from [$app_result_fname]" >&2
+# |!!>|     local valueToSet="$1"
+# |!!>|
+# |!!>|     if [[ "$valueToSet" == "read-and-flush" ]] ; then
+# |!!>|         echo "!!> ${_comment}: Flushing value file [$app_result_fname]" >&2
+# |!!>|         [[ -n "${app_result_fname:-}" ]] && rm -rf "$app_result_fname"
+# |!!>|         return "$_readValue"
+# |!!>|     fi
+# |!!>|     if [[ "$valueToSet" -gt "$_readValue" ]] ; then
+# |!!>|         echo "!!> ${_comment}: app_result_include($*) changed $_readValue -> $valueToSet" >&2
+# |!!>|
+# |!!>|         [[ -z "${app_result_fname:-}" ]] && export app_result_fname="$(mktemp "app_result_include_XXXXX.value")"
+# |!!>|         echo "$valueToSet" > "$app_result_fname"
+# |!!>|         echo "!!> ${_comment}: ----- $app_result_fname now contains: $(<"$app_result_fname")" >&2
+# |!!>|     else
+# |!!>|         echo "!!> ${_comment}: app_result_include($*) unchanged $_readValue" >&2
+# |!!>|     fi
+# |!!>|     set +x
+# |!!>| }
+
+function app_contentsFull()
+{
+    local value=0
+    {
+        local _runValue=0
+        load_params "$@" || _runValue="$?"
+         # |!!>| echo "!!> app_contentsFull.load_params() : $_runValue" >&2
+        [[ "${UKKO_VERBOSITY:-}" == 'all' ]] && app_dumpInfo "ℹ️  Summarising parameters and environment:  \$UKKO_VERBOSITY=${UKKO_VERBOSITY@Q}"
+        app_run "$@" || _runValue="$?"
+        # |!!>| echo "!!> app_contentsFull.app_run() :  $_runValue" >&2
+
+        return "$_runValue"
+    } || value=0
+
+    # |!!>| echo "!!> app_contentsFull: $value" >&2
+    # |!!>| app_result_include "$?" 'app_contentsFull'
+
+    # |!!>| app_result_include read-and-flush 'app_contentsFull'
+
+    return "$value"
+}
+
+
+source "${BUILD_FUNCS_DIR%/}/_internalUse/lib-wrapping.inc.bash"
+declare -F app_init  &> /dev/null  && app_init   # Must be outside of the 'full' as it sets up things for the tree ..
+
+_xxa=0
+# |!!>| echo "!!> lib-app Calling" >&2
+doRunWithWrapping app_contentsFull "$@" || _xxa=$?
+# |!!>| echo "!!> lib-app returned[$0] with $_xxa" >&2
+exit "$_xxa"
